corrected some changes that are done in previous commit
