+++ src/index.ts
@@ -1,9 +1,9 @@
 import { Request, Response } from "express"
-import pool from "./db/dbConfig"
 import express from "express";
 import authRoutes from "./routes/auth"
-import dotenv from "dotenv";
-dotenv.config();
+import dotenv from 'dotenv'
+dotenv.config()
+
 
 
 const app = express()
@@ -12,21 +12,8 @@ const port = 3000
 app.use(express.json())
 
 
-async function testDbConnection(){
-   try {
-       const res = await pool.query('SELECT NOW()')
-       console.log('Database is sucessfully connected ')
-       console.log(res.rows[0].now);
-   } catch (error) {
-       console.log("Error in connecting to the database")
-       console.log(error)
-   }
-}
-
-testDbConnection()
-
 app.get('/', (req:Request, res:Response) => {
-    res.status(201).json({"message": 'Hello from server'})
+    res.status(200).json({"message": 'Hello from server'})
 })
