+++ src/types/user_login.ts
@@ -1,0 +1,11 @@
+import { z } from "zod"
+import { SignupSchema } from "./user_signup"
+
+const LoginSchema = SignupSchema.pick({
+    email: true,
+    password:true
+})
+
+type UserLogin = z.infer<typeof LoginSchema>
+
+export {LoginSchema,UserLogin}
